Add dry-run and clean unused actions in som struct

+++ action.go
@@ -26,20 +26,25 @@ func action(f AnalysedFile, actions []Action) error {
 		case copyOperation:
 			dest := path.Join(action.Destination, path.Base(f.Path))
 			log.Printf("Copying file %v to %v", f.Path, dest)
-			err := copy.Copy(f.Path, dest)
-			if err != nil {
-				return err
+
+			if !dryRun {
+				err := copy.Copy(f.Path, dest)
+				if err != nil {
+					return err
+				}
 			}
 		case moveOperation:
 			dest := path.Join(action.Destination, path.Base(f.Path))
 			log.Printf("Moving file %v to %v", f.Path, dest)
 			err := copy.Copy(f.Path, dest)
-			if err != nil {
-				return err
-			}
-			err = os.Remove(f.Path)
-			if err != nil {
-				return err
+			if !dryRun {
+				if err != nil {
+					return err
+				}
+				err = os.Remove(f.Path)
+				if err != nil {
+					return err
+				}
 			}
 		default:
 			return errors.New("Error : Unknown action")
